dfsf

+++ WishList/Controllers/ItemController.cs
@@ -1,22 +1,28 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
 using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager; 
         }
 
         public IActionResult Index()
-        {
-            var model = _context.Items.ToList();
+        { 
+            var model = _context.Items.ToList().Where(UserManager.GetUserAsync(HttpContext.User));
 
             return View("Index", model);
         }
@@ -38,6 +44,9 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+
+            if (_userManager.GetUserAsync(HttpContext.User) != User.Identity)
+                return Unauthorized(); 
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
